@@ -14,7 +14,7 @@
     }
   ],
   "require": {
-    "php": ">=5.2.1",
+    "php": ">=5.4.0",
     "roundcube/plugin-installer": ">=0.1.2"
   },
   "extra": {
